Fixed paths broken by moving picker components to a separate folder
