+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/location-picker.js
@@ -4,7 +4,7 @@
 
 import angular from "angular";
 import Immutable from "immutable"; // also exports itself as (window).L
-import L from "../leaflet-bundleable.js";
+import L from "../../leaflet-bundleable.js";
 import {
   attach,
   searchNominatim,
@@ -13,10 +13,10 @@ import {
   leafletBounds,
   delay,
   getIn,
-} from "../utils.js";
-import { doneTypingBufferNg, DomCache } from "../cstm-ng-utils.js";
+} from "../../utils.js";
+import { doneTypingBufferNg, DomCache } from "../../cstm-ng-utils.js";
 
-import { initLeaflet } from "../won-utils.js";
+import { initLeaflet } from "../../won-utils.js";
 
 const serviceDependencies = ["$scope", "$element", "$sce"];
 function genComponentConf() {
