+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/details/route-picker.js
@@ -1,6 +1,6 @@
 import angular from "angular";
 import Immutable from "immutable"; // also exports itself as (window).L
-import L from "../leaflet-bundleable.js";
+import L from "../../leaflet-bundleable.js";
 import {
   attach,
   searchNominatim,
@@ -9,10 +9,10 @@ import {
   leafletBounds,
   delay,
   getIn,
-} from "../utils.js";
-import { doneTypingBufferNg, DomCache } from "../cstm-ng-utils.js";
+} from "../../utils.js";
+import { doneTypingBufferNg, DomCache } from "../../cstm-ng-utils.js";
 
-import { initLeaflet } from "../won-utils.js";
+import { initLeaflet } from "../../won-utils.js";
 
 const serviceDependencies = ["$scope", "$element", "$sce"];
 function genComponentConf() {
